requirements.txt
streamlit>=1.20
boto3
requests
moviepy==1.0.3
imageio-ffmpeg
ffmpeg-python
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pytest
pyngrok
pillow
gtts
pydub
openai==0.28.0